se refactora codigo generando nuevas clases ..se desactivan funciones erroneas...se prueba... ..
